+++ tests/test_readme.py
@@ -13,7 +13,7 @@ def test_readme(capsys: pytest.CaptureFixture, tmp_path: Path) -> None:
 
     code = README.read_text().split("```python")[1].split("```")[0]
     code = code.replace("some_file.nd2", str(SAMPLE.absolute()))
-    code = code.replace("output.some.tif", str(tmp_path / "output.some.tif"))
+    code = code.replace("output.ome.tif", str(tmp_path / "output.ome.tif"))
     exec(code)
     captured = capsys.readouterr()
     assert captured.out.startswith("Metadata(")  # because of the print statement